Tribe direct payments
On tribe direct payments, we can receive a message of type direct payment but the payment didn't come to me actually but to other member on the tribe. So the payment icon is always the SENT arrow icon (no matter if the payment is a received or a sent payment)
On the received message row of a tribe direct payment, the "sat" label should be removed and both the icon and the amount should be on grey color.