Image and Direct Payment
Images should have a fixed width (something like 25% of the chat view or so), and the bubble should fit that width (so the image filled the entire bubble). Then it will fill the needed height to keep the image ratio.
If the message has text, then text should fit the image width (not more)
Image should have no margin on left/top/right, but it should have a bigger margin to the text at the bottom

Fixed send icon issue for 1:1 Direct payment